refactor: split wechatsync bridge runtime
